requirements.txt
langgraph==1.0.4
langgraph-prebuilt==1.0.5
langgraph-checkpoint==2.1.0
langgraph-sdk==0.2.2
langchain-core==1.1.3
langchain-groq==1.1.0
langchain-community==0.4.1
pydantic==2.7.4
xxhash==3.5.0
python-dotenv
uvicorn
fastapi
streamlit